backend/requirements.txt
Flask==2.0.1
moviepy==1.0.3
Werkzeug==2.0.1
python-dotenv==0.19.0
flask-cors==3.0.10
ffmpeg-python==0.2.0 